third_party/blink/web_tests/external/wpt/css/cssom/rule-restrictions-expected.txt

This is a testharness.js-based test.
FAIL @page assert_equals: Transform doesn't quite apply to pages expected 1 but got 2
PASS @keyframe
Harness: the test ran to completion.
